Changes Neutron Security Group Ports to Integers
Previously, neutron security group tcp/udp port numbers were
defined as strings.  This was not an issue when testing with trunk
but is an issue when testing with 2014.1.b2.

2014.1.b2 expects the parameters to be defined as integers not
strings.  When the tcp/udp ports are defined as integers, the
security group rules work with trunk as well as 2014.1.b2.

Change-Id: If7e5af163993b4cad3f31583ea73e9e856ab0ccb